Tuan_20220228_20220314

+++ Back-end/BookStoreApi/Controllers/BookController.cs
@@ -3,7 +3,8 @@ using BookStoreApi.Models;
 using BookStoreApi.Services;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
-using System.Dynamic;
+using BookStoreApi.Interfaces;
+using System.Net.Http.Headers;
 
 namespace BookStoreApi.Controllers
 {
@@ -11,10 +12,33 @@ namespace BookStoreApi.Controllers
     [Route("api/[controller]")]
     public class BookController : ControllerBase
     {
-        private readonly BooksService _bookService;
-        private readonly CategoryService _categoryService;
+        private readonly IBookService _bookService;
+        private readonly ICategoryService _categoryService;
         private readonly IMapper _mapper;
-        public BookController(BooksService booksService,CategoryService categoryService,IMapper mapper)
+        private static void MergeChunks(string chunk1, string chunk2)
+        {
+            FileStream fs1 = null;
+            FileStream fs2 = null;
+            try
+            {
+                fs1 = System.IO.File.Open(chunk1, FileMode.Append);
+                fs2 = System.IO.File.Open(chunk2, FileMode.Open);
+                byte[] fs2Content = new byte[fs2.Length];
+                fs2.Read(fs2Content, 0, (int)fs2.Length);
+                fs1.Write(fs2Content, 0, (int)fs2.Length);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message + " : " + ex.StackTrace);
+            }
+            finally
+            {
+                if (fs1 != null) fs1.Close();
+                if (fs2 != null) fs2.Close();
+                System.IO.File.Delete(chunk2);
+            }
+        }
+        public BookController(IBookService booksService,ICategoryService categoryService,IMapper mapper)
         {
             _bookService = booksService;
             _categoryService = categoryService;
@@ -37,12 +61,12 @@ namespace BookStoreApi.Controllers
             return Ok(book);
         }
         [HttpPost]
-        public async Task<IActionResult> CreateNewBook([FromBody] BookDTO bookDTO)
+        public async Task<IActionResult> CreateNewBook([FromForm] BookDTO bookDTO,IFormFile File)
         {
-            Book newBook = new Book();
+            Book newBook = new Book();    
             this._mapper.Map(bookDTO, newBook);
             var findBook = await this._bookService.ValidateBook(newBook.ID,newBook.BookName);
-            if(findBook != null)
+            if(findBook != null)    
             {
                 ModelState.AddModelError("Error", "Name is exits");
                 return BadRequest(ModelState);
@@ -53,15 +77,45 @@ namespace BookStoreApi.Controllers
                 ModelState.AddModelError("Error", "Foreign key (CategoryId) does not exist");
                 return BadRequest(ModelState);
             }
+            //Upload image
+                //var formCollection = await Request?.ReadFormAsync();
+                //var image = formCollection.Files.FirstOrDefault();
+                var folderName = Path.Combine("wwwroot", "Images");
+                var pathToSave = Path.Combine(Directory.GetCurrentDirectory(), folderName);
+                if (File.Length <= 0)
+                {
+                    return BadRequest();
+                }
+                var fileName = Path.GetRandomFileName();
+                var fullPath = Path.Combine(pathToSave, fileName);
+                using (var stream = System.IO.File.Create(fullPath))
+                {
+                    await File.CopyToAsync(stream);
+                }
             findCategory.Quantity += 1;
             await this._categoryService.UpdateCategory(findCategory.Id, findCategory);
             CategoryShow category = this._mapper.Map<CategoryShow>(findCategory);
             newBook.Category = category;
+            newBook.ImagePath = fileName;
             await this._bookService.CreateAsync(newBook);
             return CreatedAtAction(nameof(GetItemBook), new { id = newBook.ID }, newBook);
         }
+        [HttpGet("image/{dbPath}")]
+        public IActionResult SeePicture(string dbPath)
+        {
+            var folderName = Path.Combine("wwwroot", "Images");
+            var pathToSave = Path.Combine(Directory.GetCurrentDirectory(), folderName);
+            var pathImage = Path.Combine(pathToSave, dbPath).Replace("/", "\\");
+            if (!System.IO.File.Exists(pathImage))
+            {
+                ModelState.AddModelError("Error", $"Could not find file {dbPath}");
+            }
+            var image = System.IO.File.OpenRead(pathImage);
+            return File(image, "image/jpeg");
+        }
         [HttpPut("detail/{id:length(24)}")]
-        public async Task<IActionResult> UpdateBookItem_id(string id, [FromBody] BookDTO updateBook)
+        [RequestFormLimits(MultipartBodyLengthLimit = 2147483648)]
+        public async Task<IActionResult> UpdateBookItem_id(string id, [FromForm] BookDTO updateBook)
         {
             var book = await this._bookService.GetAsync(id);
             if(book is null)
@@ -80,6 +134,27 @@ namespace BookStoreApi.Controllers
                 ModelState.AddModelError("Error", "Foreign key (CategoryId) does not exist");
                 return BadRequest(ModelState);
             }
+            //Upload file if file exist
+            var formRequest = await Request.ReadFormAsync();
+            var file = formRequest.Files.FirstOrDefault();
+            if(file != null)
+            {
+                var folderName = Path.Combine("wwwroot", "Images");
+                var pathToSave = Path.Combine(Directory.GetCurrentDirectory(), folderName);
+                var fileOld = Path.Combine(pathToSave, book.ImagePath).Replace("/", "\\");
+                System.IO.File.Delete(fileOld);
+                var fileNew = Path.GetRandomFileName();
+                var pathFileNew = Path.Combine(pathToSave,fileNew);
+                if(file.Length <= 0)
+                {
+                    return BadRequest();
+                }
+                using(var stream = System.IO.File.Create(pathFileNew))
+                {
+                    await file.CopyToAsync(stream);
+                }
+                book.ImagePath = fileNew;
+            }
             if(findCategory.Id != book.CategoryId) {
                 if(book.CategoryId != null)
                 {
@@ -96,120 +171,6 @@ namespace BookStoreApi.Controllers
             await this._bookService.UpdateAsync(book.ID, book);
             return CreatedAtAction(nameof(GetItemBook), new { id = book.ID}, book);
         }
-        
-        //Test API post and put
-        [HttpPut("detail")]
-        public async Task<IActionResult> UpdateBookItem_noid([FromBody] BookDTO updateBook)
-        {
-            var book = await this._bookService.GetAsync(updateBook.ID);
-            if (book is null)
-            {
-                ModelState.AddModelError("Error", "Book not found");
-            }
-            var validateBook = await this._bookService.ValidateBook(updateBook.ID, updateBook.Name);
-            if (validateBook != null)
-            {
-                ModelState.AddModelError("Error", "BookName is exist");
-                return BadRequest(ModelState);
-            }
-            Category findCategory = await this._categoryService.GetCategoryById(updateBook.CategoryId);
-            if (findCategory is null)
-            {
-                ModelState.AddModelError("Error", "Foreign key (CategoryId) does not exist");
-                return BadRequest(ModelState);
-            }
-            if (findCategory.Id != book.CategoryId)
-            {
-                if (book.CategoryId != null)
-                {
-                    Category oldCategory = await this._categoryService.GetCategoryById(book.CategoryId);
-                    oldCategory.Quantity -= 1;
-                    await this._categoryService.UpdateCategory(oldCategory.Id, oldCategory);
-                }
-                findCategory.Quantity += 1;
-                await this._categoryService.UpdateCategory(findCategory.Id, findCategory);
-            }
-            this._mapper.Map(updateBook, book);
-            CategoryShow category = this._mapper.Map<CategoryShow>(findCategory);
-            book.Category = category;
-            await this._bookService.UpdateAsync(book.ID, book);
-            return CreatedAtAction(nameof(GetItemBook), new { id = book.ID }, book);
-        }
-        [HttpPost("detailpost")]
-        public async Task<IActionResult> UpdateBookItem_noid_post([FromBody] BookDTO updateBook)
-        {
-            var book = await this._bookService.GetAsync(updateBook.ID);
-            if (book is null)
-            {
-                ModelState.AddModelError("Error", "Book not found");
-            }
-            var validateBook = await this._bookService.ValidateBook(updateBook.ID, updateBook.Name);
-            if (validateBook != null)
-            {
-                ModelState.AddModelError("Error", "BookName is exist");
-                return BadRequest(ModelState);
-            }
-            Category findCategory = await this._categoryService.GetCategoryById(updateBook.CategoryId);
-            if (findCategory is null)
-            {
-                ModelState.AddModelError("Error", "Foreign key (CategoryId) does not exist");
-                return BadRequest(ModelState);
-            }
-            if (findCategory.Id != book.CategoryId)
-            {
-                if (book.CategoryId != null)
-                {
-                    Category oldCategory = await this._categoryService.GetCategoryById(book.CategoryId);
-                    oldCategory.Quantity -= 1;
-                    await this._categoryService.UpdateCategory(oldCategory.Id, oldCategory);
-                }
-                findCategory.Quantity += 1;
-                await this._categoryService.UpdateCategory(findCategory.Id, findCategory);
-            }
-            this._mapper.Map(updateBook, book);
-            CategoryShow category = this._mapper.Map<CategoryShow>(findCategory);
-            book.Category = category;
-            await this._bookService.UpdateAsync(book.ID, book);
-            return CreatedAtAction(nameof(GetItemBook), new { id = book.ID }, book);
-        }
-        [HttpPut("detailandpayload")]
-        public async Task<ActionResult<Book>> UpdateBookItem_returnpayload([FromBody] BookDTO updateBook)
-        {
-            var book = await this._bookService.GetAsync(updateBook.ID);
-            if (book is null)
-            {
-                ModelState.AddModelError("Error", "Book not found");
-            }
-            var validateBook = await this._bookService.ValidateBook(updateBook.ID, updateBook.Name);
-            if (validateBook != null)
-            {
-                ModelState.AddModelError("Error", "BookName is exist");
-                return BadRequest(ModelState);
-            }
-            Category findCategory = await this._categoryService.GetCategoryById(updateBook.CategoryId);
-            if (findCategory is null)
-            {
-                ModelState.AddModelError("Error", "Foreign key (CategoryId) does not exist");
-                return BadRequest(ModelState);
-            }
-            if (findCategory.Id != book.CategoryId)
-            {
-                if (book.CategoryId != null)
-                {
-                    Category oldCategory = await this._categoryService.GetCategoryById(book.CategoryId);
-                    oldCategory.Quantity -= 1;
-                    await this._categoryService.UpdateCategory(oldCategory.Id, oldCategory);
-                }
-                findCategory.Quantity += 1;
-                await this._categoryService.UpdateCategory(findCategory.Id, findCategory);
-            }
-            this._mapper.Map(updateBook, book);
-            CategoryShow category = this._mapper.Map<CategoryShow>(findCategory);
-            book.Category = category;
-            await this._bookService.UpdateAsync(book.ID, book);
-            return Ok(book);
-        }
-        //
         [HttpDelete("{id:length(24)}")]
         public async Task<IActionResult> DeleteItemBook(string id)
         {
@@ -225,8 +186,15 @@ namespace BookStoreApi.Controllers
                 category.Quantity -= 1;
                 await this._categoryService.UpdateCategory(category.Id, category);
             }
-            await this._bookService.DeleteAsync(id);
-            return StatusCode(200,"Delete success");
+            if(book.ImagePath != null)
+            {
+                var folderName = Path.Combine("wwwroot", "Images");
+                var pathToSave = Path.Combine(Directory.GetCurrentDirectory(), folderName);
+                var pathToFile = Path.Combine(pathToSave, book.ImagePath).Replace("/", "\\");
+                await this._bookService.DeleteAsync(id);
+                System.IO.File.Delete(pathToFile);
+            }
+            return NoContent();
         }
         [HttpPatch("{id:length(24)}")]
         public async Task<IActionResult> UpdatePatch(string id,[FromBody] JsonPatchDocument<BookDTO> updateBook)
@@ -270,6 +238,38 @@ namespace BookStoreApi.Controllers
             findBook.Category = category;
             await this._bookService.UpdateAsync(findBook.ID, findBook);
             return CreatedAtAction(nameof(GetItemBook), new { id = findBook.ID }, findBook);
+        }
+        [HttpPost("upload")]
+        [RequestFormLimits(MultipartBodyLengthLimit = 1073741824)]
+        public async Task<IActionResult> UploadFile(IFormFile file)
+        {
+            var pathFolder = Path.Combine("wwwroot", "Temp");
+            var pathToSave = Path.Combine(Directory.GetCurrentDirectory(), pathFolder);
+            var fileName = ContentDispositionHeaderValue.Parse(file.ContentDisposition).FileName.Trim('"');
+            var pathFull = Path.Combine(pathToSave, fileName);
+            if(file.Length <= 0)
+            {
+                return BadRequest();
+            }
+            using(var stream = System.IO.File.Create(pathFull))
+            {
+                await file.CopyToAsync(stream);
+            }
+            return Ok(file);
+        }
+        [HttpPost("filemerge/{fileName}")]
+        
+        public async Task<IActionResult> FileMerge(string fileName)
+        {
+            string tempPath = Path.Combine("wwwroot","Temp");
+            string pathFolder = Path.Combine("wwwroot", "Images");
+            string newPath = Path.Combine(pathFolder, fileName);
+            string[] filePaths = Directory.GetFiles(tempPath,fileName+"*");
+            foreach (string filePath in filePaths)
+            {
+                MergeChunks(newPath, filePath);
+            }
+            return Ok(fileName);
         }
     }
 }
